backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.6.0
pydantic==2.10.6
pydantic-settings==2.8.1
email-validator==2.2.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
httpx==0.25.1
python-dotenv==1.0.0
motor==3.3.1
pandas==2.1.1
matplotlib==3.8.0
sqlparse==0.4.4
pymysql==1.1.0
aiohttp==3.8.5
